Create add category

+++ src/main/java/com/practice/web/repository/CategoryRepository.java
@@ -2,7 +2,8 @@ package com.practice.web.repository;
 
 import com.practice.web.model.Category;
 import org.springframework.data.jpa.repository.JpaRepository;
+import org.springframework.data.repository.CrudRepository;
 
-public interface CategoryRepository extends JpaRepository<Category, Integer> {
+public interface CategoryRepository extends JpaRepository<Category, Long> {
 
 }
